+++ app/views/standings/index.html.erb
@@ -1,7 +1,7 @@
 <div class="row">
   <div class="col s12">
     <table class="c bordered">
-      <h1>Players</h1>
+      <h1>Standings</h1>
       <thead>
         <tr>
           <th>Placing</th>
@@ -11,7 +11,9 @@
         </tr>
       </thead>
       <tbody id="active-players">
-        <%= render partial: "standings", locals: { players: @players} %>
+        <% @players.each do |player| %>
+        <%= render partial: "standings", locals: { player: player } %>
+        <% end %>
       </tbody>
 
     </table>
