+++ lib/pairing_tool.rb
@@ -1,16 +1,5 @@
+require "points_calculator"
 class PairingTool
-  # Group players by rank
-  ABOVE_BAR_RATING_BAND = 4
-  BAND_TWO = 1
-  BAND_THREE = -6
-  BAND_FOUR = -12
-  BAND_FIVE = -31
-  # initial tournament points
-  ABOVE_BAR_ITP = 6
-  BAND_TWO_ITP = 4
-  BAND_THREE_ITP = 3
-  BAND_FOUR_ITP = 2
-  BAND_FIVE_ITP = 0
 
   attr_reader :players, :tournament, :round
   def initialize(players:, tournament:, round:)
@@ -25,7 +14,7 @@ class PairingTool
     sorted_players = players.sort_by(&:rating)
 
     sorted_players.each do |player|
-      determine_tournament_points(player)
+      PointsCalculator.determine_tournament_points(player)
       determine_previous_opponents(player)
     end
 
@@ -36,38 +25,38 @@ class PairingTool
   end
 
   private
-
-  def initial_tournament_points(player)
-    player.tournament_points = 0
-    if player.rating >= ABOVE_BAR_RATING_BAND
-      player.tournament_points = ABOVE_BAR_ITP
-    elsif player.rating >= BAND_TWO
-      player.tournament_points = BAND_TWO_ITP
-    elsif player.rating >= BAND_THREE
-      player.tournament_points = BAND_THREE_ITP
-    elsif player.rating >= BAND_FOUR
-      player.tournament_points = BAND_FOUR_ITP
-    elsif player.rating >= BAND_FIVE
-      player.tournament_points = BAND_FIVE_ITP
-    end
-  end
-
-  def determine_tournament_points(player)
-    player.tournament_points = initial_tournament_points(player)
-    games = player.games.joins(
-      :tournament,
-    ).where(
-      tournaments: { id: tournament.id },
-    )
-    games.each do |game|
-      if game.result.winner == player
-        player.tournament_points += 1
-      end
-      if game.result.draw == true
-        player.tournament_points += 0.5
-      end
-    end
-  end
+  # extracted to points_calculator class
+  # def initial_tournament_points(player)
+  #   player.tournament_points = 0
+  #   if player.rating >= ABOVE_BAR_RATING_BAND
+  #     player.tournament_points = ABOVE_BAR_ITP
+  #   elsif player.rating >= BAND_TWO
+  #     player.tournament_points = BAND_TWO_ITP
+  #   elsif player.rating >= BAND_THREE
+  #     player.tournament_points = BAND_THREE_ITP
+  #   elsif player.rating >= BAND_FOUR
+  #     player.tournament_points = BAND_FOUR_ITP
+  #   elsif player.rating >= BAND_FIVE
+  #     player.tournament_points = BAND_FIVE_ITP
+  #   end
+  # end
+  #
+  # def determine_tournament_points(player)
+  #   player.tournament_points = initial_tournament_points(player)
+  #   games = player.games.joins(
+  #     :tournament,
+  #   ).where(
+  #     tournaments: { id: tournament.id },
+  #   )
+  #   games.each do |game|
+  #     if game.result.winner == player
+  #       player.tournament_points += 1
+  #     end
+  #     if game.result.draw == true
+  #       player.tournament_points += 0.5
+  #     end
+  #   end
+  # end
 
   def determine_previous_opponents(player)
     player.previous_opponents = []
